+++ src/service/api/user/verify-email.get.ts
@@ -8,20 +8,19 @@ import UserModel from "../../../model/mongo/user";
 import verificationCodeModel from "../../../model/mongo/verification-code";
 import { errorMessages, statusCodes } from "../../../utils/http-status";
 import { ErrorResponse, SuccessResponse } from "../../../utils/response";
+import { hasErrors } from "../../../utils/api";
 
 export const GET_VerifyEmailValidator = [body("code").exists().isString()];
 
 const GET_VerifyEmail = async (req: Request, res: Response) => {
   try {
+    if (hasErrors(req, res)) return;
     const code: string = req.query.code as string;
     const dbCode = await verificationCodeModel.findOne({ code }).exec();
-
     if (!dbCode) {
       return res.status(statusCodes.clientInputError).json(new ErrorResponse(errorMessages.clientInputError));
     }
-
     await UserModel.updateOne({ _id: dbCode.belongsTo }, { $set: { emailVerified: true } });
-
     verificationCodeModel.deleteOne({ code }).exec();
     return res
       .status(statusCodes.success)
